feat: [PL-30098]: RBAC Enforcement on Streaming Destination (#42695)
* feat: [PL-30098]: RBAC Enforcement on Streaming Destination

* feat: [PL-30098]: Tests

* fix: comments

+++ platform-service/modules/audit-service/src/main/java/io/harness/audit/remote/v1/api/streaming/StreamingDestinationsApiImpl.java
@@ -7,6 +7,14 @@
 
 package io.harness.audit.remote.v1.api.streaming;
 
+import static io.harness.audit.remote.v1.api.streaming.StreamingDestinationPermissions.DELETE_STREAMING_DESTINATION_PERMISSION;
+import static io.harness.audit.remote.v1.api.streaming.StreamingDestinationPermissions.EDIT_STREAMING_DESTINATION_PERMISSION;
+import static io.harness.audit.remote.v1.api.streaming.StreamingDestinationPermissions.VIEW_STREAMING_DESTINATION_PERMISSION;
+import static io.harness.audit.remote.v1.api.streaming.StreamingDestinationResourceTypes.STREAMING_DESTINATION;
+
+import io.harness.accesscontrol.acl.api.Resource;
+import io.harness.accesscontrol.acl.api.ResourceScope;
+import io.harness.accesscontrol.clients.AccessControlClient;
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.audit.api.streaming.StreamingService;
@@ -34,9 +42,13 @@ import org.springframework.data.domain.Pageable;
 public class StreamingDestinationsApiImpl implements StreamingDestinationsApi {
   private final StreamingService streamingService;
   private final StreamingDestinationsApiUtils streamingDestinationsApiUtils;
+  private final AccessControlClient accessControlClient;
 
   @Override
   public Response createStreamingDestinations(@Valid StreamingDestinationDTO body, String harnessAccount) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(harnessAccount, null, null),
+        Resource.of(STREAMING_DESTINATION, null), EDIT_STREAMING_DESTINATION_PERMISSION);
+
     StreamingDestination streamingDestination = streamingService.create(harnessAccount, body);
 
     log.info(String.format(
@@ -48,6 +60,9 @@ public class StreamingDestinationsApiImpl implements StreamingDestinationsApi {
 
   @Override
   public Response deleteDisabledStreamingDestination(String streamingDestinationIdentifier, String harnessAccount) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(harnessAccount, null, null),
+        Resource.of(STREAMING_DESTINATION, streamingDestinationIdentifier), DELETE_STREAMING_DESTINATION_PERMISSION);
+
     streamingService.delete(harnessAccount, streamingDestinationIdentifier);
 
     log.info(String.format(
@@ -57,6 +72,9 @@ public class StreamingDestinationsApiImpl implements StreamingDestinationsApi {
 
   @Override
   public Response getStreamingDestination(String streamingDestinationIdentifier, String harnessAccount) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(harnessAccount, null, null),
+        Resource.of(STREAMING_DESTINATION, streamingDestinationIdentifier), VIEW_STREAMING_DESTINATION_PERMISSION);
+
     StreamingDestination streamingDestination =
         streamingService.getStreamingDestination(harnessAccount, streamingDestinationIdentifier);
 
@@ -86,6 +104,9 @@ public class StreamingDestinationsApiImpl implements StreamingDestinationsApi {
   @Override
   public Response updateStreamingDestination(String streamingDestinationIdentifier,
       @Valid StreamingDestinationDTO streamingDestinationDTO, String harnessAccount) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(harnessAccount, null, null),
+        Resource.of(STREAMING_DESTINATION, streamingDestinationIdentifier), EDIT_STREAMING_DESTINATION_PERMISSION);
+
     StreamingDestination streamingDestination =
         streamingService.update(streamingDestinationIdentifier, streamingDestinationDTO, harnessAccount);
